+++ src/mkvtoolnix-gui/util/settings.cpp
@@ -343,7 +343,8 @@ Settings::load() {
   m_switchToJobOutputAfterStarting     = reg.value(s_valSwitchToJobOutputAfterStarting, false).toBool();
   m_resetJobWarningErrorCountersOnExit = reg.value(s_valResetJobWarningErrorCountersOnExit, false).toBool();
   m_removeOutputFileOnJobFailure       = reg.value(s_valRemoveOutputFileOnJobFailure,       false).toBool();
-  m_jobRemovalPolicy                   = static_cast<JobRemovalPolicy>(reg.value(s_valJobRemovalPolicy, static_cast<int>(JobRemovalPolicy::Never)).toInt());
+  m_jobRemovalPolicy                   = static_cast<JobRemovalPolicy>(reg.value(s_valJobRemovalPolicy,       static_cast<int>(JobRemovalPolicy::Never)).toInt());
+  m_jobRemovalOnExitPolicy             = static_cast<JobRemovalPolicy>(reg.value(s_valJobRemovalOnExitPolicy, static_cast<int>(JobRemovalPolicy::Never)).toInt());
   m_removeOldJobs                      = reg.value(s_valRemoveOldJobs,                                  true).toBool();
   m_removeOldJobsDays                  = reg.value(s_valRemoveOldJobsDays,                              14).toInt();
 
@@ -690,6 +691,7 @@ Settings::save()
   reg.setValue(s_valResetJobWarningErrorCountersOnExit, m_resetJobWarningErrorCountersOnExit);
   reg.setValue(s_valRemoveOutputFileOnJobFailure,       m_removeOutputFileOnJobFailure);
   reg.setValue(s_valJobRemovalPolicy,                   static_cast<int>(m_jobRemovalPolicy));
+  reg.setValue(s_valJobRemovalOnExitPolicy,             static_cast<int>(m_jobRemovalOnExitPolicy));
   reg.setValue(s_valRemoveOldJobs,                      m_removeOldJobs);
   reg.setValue(s_valRemoveOldJobsDays,                  m_removeOldJobsDays);