+++ src/mkvtoolnix-gui/util/settings.h
@@ -172,7 +172,7 @@ public:
   ScanForPlaylistsPolicy m_scanForPlaylistsPolicy;
   unsigned int m_minimumPlaylistDuration;
 
-  JobRemovalPolicy m_jobRemovalPolicy;
+  JobRemovalPolicy m_jobRemovalPolicy, m_jobRemovalOnExitPolicy;
   bool m_removeOldJobs;
   int m_removeOldJobsDays;
   bool m_useDefaultJobDescription, m_showOutputOfAllJobs, m_switchToJobOutputAfterStarting, m_resetJobWarningErrorCountersOnExit;