+++ src/mkvtoolnix-gui/util/settings_names.h
@@ -55,6 +55,7 @@ char const * const s_valHexDumps                            = "hexDumps";
 char const * const s_valHexPositions                        = "hexPositions";
 char const * const s_valHiddenColumns                       = "hiddenColumns";
 char const * const s_valJobRemovalPolicy                    = "jobRemovalPolicy";
+char const * const s_valJobRemovalOnExitPolicy              = "jobRemovalOnExitPolicy";
 char const * const s_valLastConfigDir                       = "lastConfigDir";
 char const * const s_valLastOpenDir                         = "lastOpenDir";
 char const * const s_valLastOutputDir                       = "lastOutputDir";